+++ eslint.config.mjs
@@ -6,6 +6,7 @@ const config = [
     rules: {
       'react-hooks/set-state-in-effect': 'off',
       'react-hooks/immutability': 'off',
+      'react-hooks/preserve-manual-memoization': 'off',
     },
   },
   {